Restructuring User model usage
Previously, we were extending the User model to create a custom LNF_User.
However, there is no additional functionality to doing so. This commit
now makes full utilization of the User model as part of the django
authentication package. Confirmed functionality includes signing up,
logging in to view profile page, and logging out.